feat: remove built-in polyfill
