+++ .eslintignore
@@ -1,3 +1,2 @@
 public
 dist
-src/polyfill.js
