+++ .prettierignore
@@ -1,5 +1,4 @@
 *.log
-src/polyfill.js
 demo/.dev
 public
 dist
