+++ package.json
@@ -43,9 +43,7 @@
     "test:watch": "yarn test --watch",
     "test:cov": "yarn clean:cov && yarn test --coverage",
     "build:demo": "yarn clean:dev && yarn clean:demo && rollup -c rollup/config.js --environment BUILD:demo",
-    "build-full": "BABEL_ENV=dist rollup -c rollup/config.js --environment BUILD:full",
-    "build-pure": "BABEL_ENV=dist rollup -c rollup/config.js --environment BUILD:pure",
-    "build:dist": "run-s clean:dist build-full build-pure",
+    "build:dist": "yarn clean:dist && BABEL_ENV=dist rollup -c rollup/config.js --environment BUILD:dist",
     "build": "run-s build:*",
     "preversion": "run-s lint test build:dist",
     "postversion": "git push --follow-tags --no-verify && npm publish && yarn clean:dist",
@@ -114,8 +112,7 @@
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-standard": "^20.0.0",
     "ts-jest": "^26.4.4",
-    "typescript": "^4.1.2",
-    "web-animations-js": "^2.3.2"
+    "typescript": "^4.1.2"
   },
   "peerDependencies": {
     "react": ">= 16.8.0"
