--- src/pure.ts
@@ -1,2 +1,0 @@
-export { default } from "./useWebAnimations";
-export * from "./animations";
